Update k_pontic_steppe.txt
I'm going to keep this the same. Like I commented on the push the main reason for CK3 and CK2 vanilla to have it all as gothic was since their map was just not as detailed. I made sure to add the extra provinces into Crimea to match HIP so that the culture setup was like HIP and EU4.